Remove dependency on RestClient for HTTP requests
RestClient depends on the `ruby-unf` library which fails to compile
on OS X 10.10.3. To circumvent this issue and to avoid issues with
dependencies in the future, use Net::HTTP for sending requests since
it is part of the standard library.

The one issue with Net::HTTP is that it doesn't support
[file upload][1] very well. The advice in [this][2] Stack Overflow
post suggests using the [multipart-post library][3] instead which
extends Net::HTTP without any other dependencies.

[1]: http://ruby-doc.org/stdlib-2.2.1/libdoc/net/http/rdoc/Net/HTTP.html#class-Net::HTTP-label-POST
[2]: http://stackoverflow.com/questions/184178/ruby-how-to-post-a-file-via-http-as-multipart-form-data
[3]: https://github.com/nicksieger/multipart-post